Finished preliminary code review on AllBuds component
